Generate report from scripts referred to or embedded in a report config
Change the report executor API for it to return a writable content instead
of simply content.

Update the status as we progress and accordingly. 

Now there is no scope for resetting a event status but that will be available 
in near future. We will also have to have a way to post manual events.

Immediate next task is to set the report fields that are from 'Report' abstract
content.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>