Fill report with standard report fields
The standard report fields are -
 # Report schedule date-time
 # Execution start date-time
 # Execution end date-time
 # Parameters the report was created with
 # The configuration report created against

Only the in case of report parameters it will be checked whether its
been set or not, if not set then it will be set else ignored. This is so
that the report can fill it as it sees necessary. As there could be other
report parameters which are calculated parameters but might need
saving.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>